Move radio_frequency entity classes to entity module
