+++ homeassistant/components/radio_frequency/__init__.py
@@ -1,25 +1,23 @@
 """Provides functionality to interact with radio frequency devices."""
 
-from abc import abstractmethod
 from datetime import timedelta
 import logging
-from typing import final
 
 from rf_protocols import ModulationType, RadioFrequencyCommand
 
 from homeassistant.config_entries import ConfigEntry
-from homeassistant.const import STATE_UNAVAILABLE
 from homeassistant.core import Context, HomeAssistant, callback
 from homeassistant.exceptions import HomeAssistantError
 from homeassistant.helpers import config_validation as cv, entity_registry as er
-from homeassistant.helpers.entity import EntityDescription
 from homeassistant.helpers.entity_component import EntityComponent
-from homeassistant.helpers.restore_state import RestoreEntity
 from homeassistant.helpers.typing import ConfigType
-from homeassistant.util import dt as dt_util
 from homeassistant.util.hass_dict import HassKey
 
 from .const import DOMAIN
+from .entity import (
+    RadioFrequencyTransmitterEntity,
+    RadioFrequencyTransmitterEntityDescription,
+)
 
 __all__ = [
     "DOMAIN",
@@ -154,73 +152,3 @@ async def async_send_command(
         entity.async_set_context(context)
 
     await entity.async_send_command_internal(command)
-
-
-class RadioFrequencyTransmitterEntityDescription(
-    EntityDescription, frozen_or_thawed=True
-):
-    """Describes radio frequency transmitter entities."""
-
-
-class RadioFrequencyTransmitterEntity(RestoreEntity):
-    """Base class for radio frequency transmitter entities."""
-
-    entity_description: RadioFrequencyTransmitterEntityDescription
-    _attr_should_poll = False
-    _attr_state: None = None
-
-    __last_command_sent: str | None = None
-
-    @property
-    @abstractmethod
-    def supported_frequency_ranges(self) -> list[tuple[int, int]]:
-        """Return list of (min_hz, max_hz) tuples."""
-
-    @callback
-    @final
-    def supports_frequency(self, frequency: int) -> bool:
-        """Return whether the transmitter supports the given frequency."""
-        return any(
-            low <= frequency <= high for low, high in self.supported_frequency_ranges
-        )
-
-    @callback
-    @final
-    def supports_modulation(self, modulation: ModulationType) -> bool:
-        """Return whether the transmitter supports the given modulation."""
-        return modulation == ModulationType.OOK
-
-    @property
-    @final
-    def state(self) -> str | None:
-        """Return the entity state."""
-        return self.__last_command_sent
-
-    @final
-    async def async_send_command_internal(self, command: RadioFrequencyCommand) -> None:
-        """Send an RF command and update state.
-
-        Should not be overridden, handles setting last sent timestamp.
-        """
-        await self.async_send_command(command)
-        self.__last_command_sent = dt_util.utcnow().isoformat(timespec="milliseconds")
-        self.async_write_ha_state()
-
-    @final
-    async def async_internal_added_to_hass(self) -> None:
-        """Call when the radio frequency entity is added to hass."""
-        await super().async_internal_added_to_hass()
-        state = await self.async_get_last_state()
-        if state is not None and state.state not in (STATE_UNAVAILABLE, None):
-            self.__last_command_sent = state.state
-
-    @abstractmethod
-    async def async_send_command(self, command: RadioFrequencyCommand) -> None:
-        """Send an RF command.
-
-        Args:
-            command: The RF command to send.
-
-        Raises:
-            HomeAssistantError: If transmission fails.
-        """
